@@ -8,6 +8,7 @@ import { IMediaWithTweetId } from '@type/app.type';
 import { initMediaFromUrl } from '@utils/helper';
 import React, { memo, useMemo } from 'react';
 import InfiniteScroll from 'react-infinite-scroll-component';
+import Skeleton from 'react-loading-skeleton';
 import Masonry from 'react-masonry-css';
 import { useParams } from 'react-router';
 import { useTweetService } from 'services/tweet.service';
@@ -22,6 +23,7 @@ const ProfileMediaPage = () => {
     data: tweets,
     hasMore,
     fetchNextPage,
+    isLoading,
   } = useInfinityList({
     queryConfig: {
       limit: DEFAULT_LIST_LIMIT,
@@ -42,6 +44,32 @@ const ProfileMediaPage = () => {
     }, []);
   }, [tweets]);
 
+  const mediaSkeletons = useMemo(() => {
+    return [...Array(6)].map((_, idx: number) => {
+      return (
+        <Skeleton
+          height={Math.floor(Math.random() * 500)}
+          key={`tweet-media-skeleton-${idx}`}
+          style={{
+            marginBottom: '1rem',
+          }}
+        />
+      );
+    });
+  }, []);
+
+  if (isLoading) {
+    return (
+      <Masonry
+        className={classes.grid}
+        columnClassName={classes.column}
+        breakpointCols={MASONRY_CONFIG_BREAKPOINTS}
+      >
+        {mediaSkeletons}
+      </Masonry>
+    );
+  }
+
   return (
     <InfiniteScroll
       dataLength={medias.length}
@@ -55,7 +83,10 @@ const ProfileMediaPage = () => {
         breakpointCols={MASONRY_CONFIG_BREAKPOINTS}
       >
         {medias?.map((tweet: any) => (
-          <MediaCard data={tweet} key={`infinity-media-list-${tweet._id}`} />
+          <MediaCard
+            data={tweet}
+            key={`infinity-media-list-${tweet?.tweetId}`}
+          />
         ))}
       </Masonry>
     </InfiniteScroll>
